+++ src/core/server/utility/index.ts
@@ -2,5 +2,6 @@ export { default as closest } from './closest';
 export { default as config } from './config';
 export { default as hash } from './hash';
 export { default as math } from './math';
+export * as restrict from './restrict';
 export { default as vector } from '@AthenaShared/utility/vector';
 export { UID as uid } from '@AthenaShared/utility/uid';